pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "archivist"
version = "2.0.0"
description = "CLI tool for managing Obsidian vault frontmatter and archive workflows"
requires-python = ">=3.10"
dependencies = [
    "pyyaml",
    "argcomplete",
]

[project.scripts]
archivist = "archivist.cli:main"

[project.optional-dependencies]
dev = ["pytest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: requires git on PATH and real filesystem ops"]
pythonpath = ["."]

[tool.setuptools.packages]
find = {}

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pyright]
ignore = ["tests"]
reportUndefinedVariable = "error"